#136900 - refactor subscriptions to be in their own class by provider combine contact and timeline updates into one method, add in rsvp status to ticket data, fix fatal error if no venue.
